PR B refactor: split sourceStatus into audit/{api,store,status} + UI
The original sourceStatus.tsx was reasonable when audit data was
synchronous + bundled. Once a real network call landed, the file
straddled three concerns it shouldn't: I/O, domain logic, presentation.
Splitting now while there's only one endpoint is cheap; doing it after
PR D adds a second endpoint would be messier because callers will have
already coupled to whichever shape they found.

New layout:

  src/lib/audit/
    api.ts      types + fetch — the network contract.
    store.ts    module-level Map + listener Set + useSyncExternalStore
                hook + prefetchStatus. Depends on api.ts.
    status.ts   pure logic — REVIEWS, BROKEN_STATUS_CODES,
                STALENESS_*, isOverdue, getStatusKind, types.
                No React, no fetch.

  src/components/audit/
    StatusDot.tsx    pure-presentation dot + tooltip.
    ReportFlag.tsx   pure-presentation report-flag link.

Importers (Sources.tsx, ui.tsx, DesignLab.tsx, main.tsx) updated to
the new paths.

src/lib/sourceStatus.tsx deleted — no behaviour changes.
